refactor: move departure presenter into map presenter
